作者、小说支持分页查询，并且修改全局异常处理

+++ src/main/java/com/example/backend/entity/Novel.java
@@ -27,7 +27,7 @@ public class Novel {
     String desc;
 
     @TableField("volume")
-    int volume;
+    double volume;
 
     @TableField("page")
     int page;
